Refactoring to EMP project from template
