+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/EquinityHistoricalData.java
@@ -1,13 +1,13 @@
-package uk.gov.justice.laa.crime.microservice;
+package uk.gov.justice.laa.crime.equinity.historicaldata;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MicroserviceTemplateApplication {
+public class EquinityHistoricalData {
 
 	public static void main(String[] args) {
-		SpringApplication.run(MicroserviceTemplateApplication.class, args);
+		SpringApplication.run(EquinityHistoricalData.class, args);
 	}
 
 }
