--- microservice-template/src/main/resources/application.yaml
@@ -1,1 +1,0 @@
-
